Added new string.ts file to encapsulate the string arbitraries logic.
