+++ src/arbitraries/index.ts
@@ -1,4 +1,3 @@
-import * as utils from './util'
 import {
   Arbitrary,
   ArbitraryArray,
@@ -14,6 +13,7 @@ import {
 
 export * from './types'
 export {Arbitrary} from './internal'
+export {char, hex, base64, ascii, unicode, string,} from './string'
 
 export const integer = (min = Number.MIN_SAFE_INTEGER, max = Number.MAX_SAFE_INTEGER): Arbitrary<number> =>
   min > max ? NoArbitrary : (min === max ? new ArbitraryConstant(min) : new ArbitraryInteger(min, max))
@@ -23,35 +23,6 @@ export const real = (min = Number.MIN_SAFE_INTEGER, max = Number.MAX_SAFE_INTEGE
 
 export const nat = (min = 0, max = Number.MAX_SAFE_INTEGER): Arbitrary<number> =>
   max < 0 ? NoArbitrary : integer(Math.max(0, min), max)
-
-const charArb = (min = 0x20, max = 0x7e): Arbitrary<string> => min < 0x20 || max > 0x7e ?
-  NoArbitrary : new ArbitraryInteger(min, max).map((v) => String.fromCodePoint(v))
-
-export const char = (start?: string, end?: string): Arbitrary<string> => start === end ?
-  (start === undefined ? charArb() : charArb(start.charCodeAt(0), start.charCodeAt(0))) :
-  (end === undefined ?
-    charArb(start!.charCodeAt(0), start!.charCodeAt(0)) :
-    charArb(start!.charCodeAt(0), end.charCodeAt(0)))
-
-export const ascii = (): Arbitrary<string> =>
-  new ArbitraryInteger(0x00, 0x7f).map((v) => String.fromCodePoint(utils.printableCharactersMapper(v)))
-
-export const hex = (): Arbitrary<string> =>
-  new ArbitraryInteger(0, 15).map((v) => String.fromCodePoint(v < 10 ? v + 48 : v + 97 - 10))
-
-export const base64 = (): Arbitrary<string> =>
-  new ArbitraryInteger(0, 63).map((v) => String.fromCodePoint(utils.base64Mapper(v)))
-
-export const unicode = (encoding = 'utf-8'): Arbitrary<string> => encoding === 'utf-16' ?
-  new ArbitraryInteger(0x0000, 0x10ffff).map((v) => String.fromCodePoint(utils.printableCharactersMapper(v))) :
-  new ArbitraryInteger(0x0000, 0x10f7ff).map((v) => String.fromCodePoint(utils.utf8Mapper(v)))
-
-export const base64String = (unscaledMin = 4, unscaledMax = 16): Arbitrary<string> =>
-  string(unscaledMin + 3 - ((unscaledMin + 3) % 4), unscaledMax - (unscaledMax % 4), base64())
-    .map((s) => [s, s.slice(1), `${s}==`, `${s}=`][s.length % 4])
-
-export const string = (min = 2, max = 10, charArb = char()): Arbitrary<string> =>
-  min === 0 && max === 0 ? constant('') : array(charArb, min, max).map(a => a.join(''))
 
 export const array = <A>(arbitrary: Arbitrary<A>, min = 0, max = 10): Arbitrary<A[]> =>
   min > max ? NoArbitrary : new ArbitraryArray(arbitrary, min, max)
